# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_api_handlers.ErrorResponse:
    properties:
      code:
        type: string
      details:
        type: string
      error:
        type: string
    type: object
  internal_api_handlers.GenerateRequest:
    properties:
      analysis_type:
        type: string
      parameters:
        additionalProperties: true
        type: object
      portfolio_data:
        $ref: '#/definitions/internal_api_handlers.PortfolioSnapshot'
      type:
        description: '"weekly_summary" or "analysis"'
        type: string
      user_id:
        type: string
      week_start:
        type: string
    required:
    - portfolio_data
    - type
    - user_id
    type: object
  internal_api_handlers.GenerateResponse:
    properties:
      artifact_uri:
        type: string
      content:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      error:
        type: string
      model:
        type: string
      processing_time:
        type: string
      request_id:
        type: string
      success:
        type: boolean
      tokens_used:
        type: integer
    type: object
...
  internal_api_handlers.PortfolioSnapshot:
    properties:
      day_change:
        type: number
      day_change_pct:
        type: number
      total_return:
        type: number
      total_return_pct:
        type: number
      total_value:
        type: number
      week_change:
        type: number
      week_change_pct:
        type: number
    type: object
...
  internal_api_handlers.StoreRequest:
    properties:
      data:
        description: Base64 encoded data
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      namespace:
        type: string
      user_id:
        type: string
    required:
    - data
    - namespace
    type: object
  internal_api_handlers.StoreResponse:
    properties:
      content_hash:
        type: string
      error:
        type: string
      size:
        type: integer
      stored_at:
        type: string
      success:
        type: boolean
      uri:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@stackservice.com
    name: API Support
    url: http://www.stackservice.com/support
  description: GenZ Web3 Multi-Chain Investment Platform API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Stack Service API
  version: "1.0"
paths:
  /_internal/0g/generate:
    post:
      consumes:
      - application/json
      description: Performs AI inference for weekly summaries or on-demand analysis
      parameters:
      - description: Generation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.GenerateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.GenerateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
      summary: Generate AI analysis
      tags:
      - internal
...
  /_internal/0g/store:
    post:
      consumes:
      - application/json
      description: Stores data in 0G storage with specified namespace and metadata
      parameters:
      - description: Storage request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.StoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.StoreResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
      summary: Store data in 0G storage
      tags:
      - internal
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"